Uploading columns of deposit_insurance_database_july2015
Uploading columns of deposit_insurance_database_july2015 for the first time. The full excel file is from this link: https://datacatalog.worldbank.org/search/dataset/0040209. Any future updates should make a local copy of the existing table and upload a file with the exact same name and structure to not interrupt downstream uses.